Refactor cli/config logic
